Able to edit values of newly created products
It wasn't possible to create a product and then edit its values before
without reloading the page so the cells got assigned with a product-id.
That is fixed, and also it should also be possible to edit the cells of
products which creation is still pending, because an added queue-system.